# Fix some scaling issues with the global queue in the thread pool