@kouvel 2022-05-16
ASP.NET perf on an 80-proc arm64 machine below. For this comparison I included https://github.com/dotnet/aspnetcore/pull/40476 on both the before and after sides because in some cases the thread pool global queue bottleneck was only showing up after that fix.

| | Before | After | Diff |
|-:|-:|-:|-:|
| PlaintextPlatform | 7896486 | 11830164 | 49.8% |
| JsonPlatform | 310950 | 1146701 | 268.8% |
| FortunesPlatform | 334955 | 437976 | 30.8% |
| Plaintext | 9785381 | 10187365 | 4.1% |
| Json | 270047 | 1089592 | 303.5% |
| Fortunes | 244937 | 331054 | 35.2% |

No significant change on a 48-proc x64 machine:

| | Before | After | Diff |
|-:|-:|-:|-:|
| PlaintextPlatform | 12042603 | 12051597 | 0.1% |
| JsonPlatform | 1352148 | 1348159 | -0.3% |
| FortunesPlatform | 470702 | 478676 | 1.7% |
| Plaintext | 7760305 | 7835156 | 1.0% |
| Json | 1053096 | 1059431 | 0.6% |
| Fortunes | 397688 | 400052 | 0.6% |

No significant change on other citrine machines where only one queue is used.