@JulieLeeMSFT 2022-05-16
> No significant change on a 48-proc x64 Linux machine:

@kouvel is this RPS measurement?