@kouvel 2022-05-16
> And these numbers are without solving the same issue in aspnet?

For the ampere machine I included your fix from https://github.com/dotnet/aspnetcore/pull/40476 in the before and after columns so the numbers include the aspnet fix, as in some tests the thread pool scaling issue was only showing up after that. For the other machines I didn't include the aspnet fix.

> @kouvel is this RPS measurement?

Yes, these are RPS numbers. I'll check the mean latency on the ampere machine and will add those above.