@kouvel 2022-05-16
Here are the mean latency numbers for ASP.NET perf on the 80-proc arm64 Linux machine:

| Latency | Before (ms) | After (ms) | Diff (ms) |
|-:|-:|-:|-:|
| PlaintextPlatform | 1.43 | 1.64 | 0.21 |
| JsonPlatform | 1.86 | 0.56 | -1.30 |
| FortunesPlatform | 1.86 | 1.61 | -0.25 |
| Plaintext | 0.33 | 0.31 | -0.02 |
| Json | 0.97 | 0.31 | -0.66 |
| Fortunes | 1.52 | 1.18 | -0.34 |