@EgorBo 2022-06-16
Improvements on Linux-arm64: https://github.com/dotnet/perf-autofiling-issues/issues/6172