@kouvel 2022-07-06
There's some more info about the issue I ran into in https://github.com/dotnet/aspnetcore/issues/41391. There are more things to investigate, including polling for IO on worker threads. It is to be determined what other changes may help and if it would work better than currently. We can chat if you'd like.